pyproject.toml
[project]
name = "mwaa-kafka-project"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=2.10.5",
    "black>=25.1.0",
    "boto3>=1.37.8",
    "clickhouse-connect>=0.8.15",
    "confluent-kafka>=2.8.2",
    "faker>=36.2.2",
    "orjson>=3.10.15",
    "pandas>=2.2.3",
    "ruff>=0.9.9",
]
